pyproject.toml
[project]
name = "dbt-switch"
version = "0.2.6"
description = "A simple CLI to switch between dbt Cloud projects and hosts."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.8",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
]

[project.scripts]
dbt-switch = "dbt_switch.main:main"

[dependency-groups]
dev = [
    "pre-commit>=3.5.0",
    "pytest>=8.3.5",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.12.11",
]

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "src"}
